prepared ListViewController to pass selected travel title and ID
I made preparations in the ListViewController class to pass the selected travel title and ID to the ViewController. I added chosenTitle and chosenTitleId variables to store the selected title and ID. In the addButtonClicked() function, chosenTitle is reset before performing the segue. With the didSelectRowAt function in the tableView, the selected title and ID are determined and prepared to be passed to the ViewController when segueing.